move Fly-Src stuff to its own package
